Use guestmount to avoid having to use superuser privileges
Executing anything with superuser privileges is scary. Simple bugs can
have devastating consequences for the user. Users that have access to
the guestmount binary from guestfs/guestfish and fuse can now mount the
filesystem through it without elevated privileges. If the script is run
as root, a normal mount is done.