Use EXIT trap to cleanup and avoid rm -rf
If the script is aborted or interrupted, temporary files should be
cleaned up. To catch all possible error conditions we use a trap on EXIT
to call the cleanup function. This way, the cleanup function is called
for manual "exit" calls, when executables fail, as well as  when the script
ends successfully.

Also, instead of running "rm -rf", which is scary especially when the
script is run as root, we clean up the temporary files individually as
needed.