Avoid using pushd and popd
Any additional state adds confusion. If pushd and popd are used, then it
not clear in which directory a given line of code is executed. To make
sure, one first has to search whether pushd was used before or not.

If the current working directory changes, then the relative paths as it
is stored in variables will not refer to the same directory. But
relative paths are more clear if all relative paths throughout the
script are relative to the same directory. This is not possible if pushd
and popd are used.

Less lines of code are used if pushd and popd are avoided and options
like -C for tar are used instead.

pushd and popd are bashisms and thus their use makes it harder to port
scripts or snippets from the scripts to POSIX compliant shells.